MIG-refacto: put exceptions and entities in separate directories
